recompiler_new: signext fix (attempt)
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@14956 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/recompiler_new/cpu-all.h
@@ -255,10 +255,10 @@ void     remR3GrowDynRange(unsigned long physaddr);
 #endif
 
 void     remR3PhysRead(RTGCPHYS SrcGCPhys, void *pvDst, unsigned cb);
-uint8_t  remR3PhysReadU8(RTGCPHYS SrcGCPhys);
-int8_t   remR3PhysReadS8(RTGCPHYS SrcGCPhys);
-uint16_t remR3PhysReadU16(RTGCPHYS SrcGCPhys);
-int16_t  remR3PhysReadS16(RTGCPHYS SrcGCPhys);
+uint32_t remR3PhysReadU8(RTGCPHYS SrcGCPhys);
+int32_t  remR3PhysReadS8(RTGCPHYS SrcGCPhys);
+uint32_t remR3PhysReadU16(RTGCPHYS SrcGCPhys);
+int32_t  remR3PhysReadS16(RTGCPHYS SrcGCPhys);
 uint32_t remR3PhysReadU32(RTGCPHYS SrcGCPhys);
 int32_t  remR3PhysReadS32(RTGCPHYS SrcGCPhys);
 uint64_t remR3PhysReadU64(RTGCPHYS SrcGCPhys);
